rembg-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
rembg>=2.0.62
pillow>=10.4.0
python-multipart>=0.0.9
